Remove use of deprecated AsyncTask
